tests: avoid intersphinx processing in legacy pages test
When attempting to sanity check the handling of legacy pages (for
cleanup), the mocked publisher does not aim to support asset-related
content. Since intersphinx database can be published by default, the
logic an attempt to retrieve page information for the intersphinx
database to publish. To avoid this, explicitly disable intersphinx
processing for this test.

Signed-off-by: James Knight <git@jdknight.me>