env: maintain tracked page identifier cache between builds
When a cache is kept for tracking previously published pages and their
identifiers, if a processing event results in pages not needing to be
updated, unchanged pages can cause their page identifiers to be removed
from the active list. This can result in a following pass to cleanup
legacy-flagged pages for pages that are still considered unchanged.

The cache implementation should be maintaining the assumed cached
identifier for pages if they remain changed for a run. Updating the
implementation to always populate the cached page identifiers before
replacing them with new active page identifiers (if any).

Signed-off-by: James Knight <git@jdknight.me>