util: handle stty failures under mingw for getpass workaround
When operating in a MinGW terminal without invoking `winpty`, getpass
does not function as expected. This extension included a workaround by
using `stty`; however, it has been observed in recent environments that
stty may fail to run. This would result in an exception and break a
user's attempt to perform a publish request.

First, this commit improves the handling when stty fails to invoke as
expected (instead of throwing exceptions a user may not understand).
Second, if stty, instead of falling back to trying `getpass` (which
seems to always never work in this environment), just stop and report
the issue. Users can should always be able to run with `winpty` to
prevent issues when using the ask-password capability.

Signed-off-by: James Knight <git@jdknight.me>