support property retry events for extension-specific properties
This extension will update page properties under a get-then-update
process, where a retry attempt will be made to help deal with conflict
scenarios. This only applied to Confluence-specific properties and not
this extension's custom property (e.g. used for page hash tracking).

This commit is updated the implementation to perform the get-then-update
with retries for all property submissions, in attempt to prevent odd
publish scenarios for users.

Signed-off-by: James Knight <git@jdknight.me>