publisher: drop space key check on connection
This extension performs a series of little checks when starting a first
connection to ensure users are indeed talking to a Confluence instance.
This is to help rule our misconfigured sites or dealing with proxy
issues. One check performed used the configured key value and sanity
checked it against what Confluence reported the key was as well. While
functional for some time, Confluence Cloud now appears to replace the
key value from a reported space key to a new system key. This causes
the checks to fail, preventing publishing. To help alleviate the issue,
we will not longer perform a key check.

Signed-off-by: James Knight <git@jdknight.me>