rest: adjust publish retry events based on transient errors
The curl documentation gives a good capture of transient errors and what
it's `--retry` call supports a retry on. Instead of having this
extension retry on all 500 errors, narrow down the scope to specific
errors more so valid for a retry. We also extend the retry events on
select timeout-related events.

Signed-off-by: James Knight <git@jdknight.me>