transmute: enforce `render_dot` file to `str` type
Old `render_dot` returned a `str` but Sphinx v8.2 comes with a wrapper
to prepare for the type changing to a `Path` type. Sphinx's usage of
`render_dot` will generate a file type and use it in the translator, but
this extension replaces the node here with a compatible `nodes.image`
type ahead of time (e.g. for asset processing). The `nodes.image` type
expects a `str` type, so convert ensure any new Path-like type is
converted back to a `str` to prevent issues.

Signed-off-by: James Knight <git@jdknight.me>