translator: support for collapsible admonitions
The following attempts to provide support for Sphinx v8.2's new
`collapsible` option for admonitions. This implementation wraps the
contents of an admonition block into an expand macro to handle this
hint.

However, this is not pretty on Confluence (v1 and v2 editors). Not much
we can do for styling approaches as we cannot manipulate macro-generated
content.

Signed-off-by: James Knight <git@jdknight.me>