rest: relax content-type check for dumping request data
When configured to log request data to the standard output stream, the
check used to consider printing JSON data checked if the content type
was marked as `application/json;charset=utf-8`. However, it has been
observed in recent testing that content type used for JSON-related
requests was set to `application/json`.

Adjusting to check to just look for `application/json` in the content
type before attempting to parse/output JSON content in a debug mode.

Signed-off-by: James Knight <git@jdknight.me>