publisher: rework default handling with property fetch call
Instead of passing a `default` value into a `get_page_property` call,
internally populate a property so that the call always returns either
a remote property value or a pre-populated property value that can be
used for requesting a new property.

All existing implementations all utilize a default value, so handling an
unset value is not required. Although, if future calls want to handle
such a scenario, they can also check the `value` option for a `None`
value.

Signed-off-by: James Knight <git@jdknight.me>