publisher: warn if multiple same-key properties are detected on a page
In theory, there should only be a mapping of one property per key on a
page. It has been reported in some scenarios that duplicate keys may
exist on a Confluence database, which appears to be an undesired state.
To help with debugging, if multiple results are detected, report this
as a warning.

Signed-off-by: James Knight <git@jdknight.me>