publisher: attempt to perform a property refresh on hidden conflict
Users have reported scenarios where a property update may fail due to a
duplicate property entry registered in Confluence's db. In this
scenario, Confluence can report a 400 error. To try to help push the
database back into an ideal state, we will request to remove the
property to cleanup then re-push a new property entry back in -- this
should be fine for all property types that this extension intends to
manage.

See-also: sphinx-contrib/confluencebuilder#1094
Signed-off-by: James Knight <git@jdknight.me>