initial/limited inlined comment check support
Provides a crude check to warn if a page update will result in the lost
of inlined comments on a page. Users can fail a publish attempt in these
situations using `--fail-on-warning` [1], if desired.

This commit also adds the `confluence_publish_skip_commented_pages`
option, to allow skipping page updates on inlined commented pages
(instead of overwriting the page).

Note that resolved comments are still considered a flagged state for
inlined comment. So, a page update will warn even if all comments are
resolved.

[1]: https://www.sphinx-doc.org/en/master/man/sphinx-build.html#cmdoption-sphinx-build-W

Signed-off-by: James Knight <git@jdknight.me>