publisher: improve error message on update attempts on restricted pages
When a user attempts to publish to Confluence that results in a page
update to a restricted page, Confluence API reports a 404 error as the
page cannot be found, versus a 403 error. This occurs when another user
may have setup view/edit permissions on a page. There is also the fun
case where if a user archives the page, the permissions are still
configured making it more confusing that the page appears to be legit
missing.

This commit attempts to improve the error handling for such scenarios.
For cases where we have found a page and failed to update with a 404,
we will assume that the page is an edit-locked page (granted, it may be
an odd timing error of someone deleting the page between fetch and
update, but we have tailored the error message to hint to "try again"
for such scenarios). In the case where a new page is being added since
we detected no existing page and a title conflict exists, we will
assume the page is a view-restricted page.

Signed-off-by: James Knight <git@jdknight.me>