adjust configuration default/environments logic to focus on app
We crudely attach our internal configuration manager on Sphinx's
application instance to utilize it in different locations where
environment overrides and configuration defaults can be applied.

However, Sphinx had deprecated the reference of the application from
builders. To help manage this, this commit adjusts the defaults and
environment override calls to be based on a provided application.

While this works fine for most use cases, the default builder case
handle this by keeping an internal reference to the application context.
This works for now and most likely will work when `Builder.app` is
removed, but we may experience issues later with this approach (unknown
at this time).

Signed-off-by: James Knight <git@jdknight.me>