transmute: handle prospect render_math signature changes
The development version of Sphinx has tweaked its `render_math` call
to now accept a configuration option. To support this, we will adjust
the use of this call to forward the configuration in. If a TypeError
is generated, we will assume this is an older version of Sphinx and try
the original signature call.

See also: https://github.com/sphinx-doc/sphinx/pull/13626

Signed-off-by: James Knight <git@jdknight.me>