builder: do not encode emoji/other unicode characters for anchor targets
We originally encoded specific characters to prevent Confluence from
suppressing anchors for select characters, but it is unknown the
extensive list of characters Confluence was not happy with. We then
switch to `quote` which worked for the most part, but when users used
Emoji's, these characters would become encoded and generate anchor
targets with incorrect values. Now, we do a partial quote in an attempt
to be flexible -- we quote the standard ASCII range using Python default
safe sets and anything beyond it, we will just leave as is.

Signed-off-by: James Knight <git@jdknight.me>