translator: delayed anchor injection for other refids
Expands the populated anchor set to include anchors for other reference
identifier not processed. To support these anchor points, we populate
known anchors to be added at a point of time where they can be added
into the page. If we do not, it will generate anchors in locations
where Confluence will render newlines. For example, adding an anchor
before a paragraph, Confluence styling will ensure the paragraph starts
on a new block leaving a less than ideal rendering. Instead, we populate
anchors as we process and then jump them into the first block we can.

Signed-off-by: James Knight <git@jdknight.me>