ignore false-positive in hyperlink checks
It appears that mypy is now generating false-positives for some of the
hyperlink check implementation. The types from the find call should all
be node types checked by `is_confluence_href_node`, which are, for
example, indexable; although mypy reports otherwise. If we replicate
the same checks inside the loop without the wrap of
`is_confluence_href_node`, mypy does not complain. For now, just ignore
the warnings.

Signed-off-by: James Knight <git@jdknight.me>