doc: move trademark hints to introduction section
A while back, LaTeX documentation was updated to generate new pages for
sections. This resulted in the introduction section from splitting up
content in a way where trademark notes were presented in an
"Asking for help" page instead of a desired introduction page. Moving
this content into the initial introduction page.

Signed-off-by: James Knight <git@jdknight.me>