rework assets to always be standalone
This extension originally introduced the concept of "shared assets" by
allowing pages to share duplicate assets by publishing an asset in the
main document and referencing the attachment from the hosted document.
While the can reduce the amount of published assets, it came with a
higher maintenance cost in the code. The code has started to become more
complex, especially when trying to introduce lazy-asset initialization
while supporting POSIX parallel builds (to be addressed in the future).
To help simplify things, if a page requires an attachment (e.g. image),
it will now always host said asset (as if the legacy configuration
option `confluence_asset_force_standalone` was enabled).

A side benefit from this change is that there should no longer be any
complications of moving/replacing the main document on Confluence
without having to deal with linked attachments.

Signed-off-by: James Knight <git@jdknight.me>