avoid duplicate asset registration for late asset registrations
A previous commit [1] did some rework in asset processing. An issue from
this change is that assets would be processed twice. This was due to
assets being first registered when first processed and then re-added
during new assets from a fetch event. This commit adjusts to only
register the internal asset when a node is first processed.

[1]: 15877c29d1355f617598d2cebeea9013fb55a149

Signed-off-by: James Knight <git@jdknight.me>