delayed asset management for multiprocessing environments
While this extension aims to be parallel safe for both reads and writes,
a limitation existed for assets (e.g. images) that are populated in the
translator's phase. While these nodes would be processed for a given
documentation run, assets would not be properly tracked for parallel
builds in POSIX environments. While late assets were detected/tracked in
the extension's asset manager, they would only be tracked for an
individual process. This resulted in only the first document's assets to
be tracked (since Sphinx internally does not farm out the initial
document).

With the aim to be flexible no matter when assets are prepared, this
extension will now track delayed assets into a multiprocessing-safe
list. If a parallel write is detected, a multiprocessing-safe list will
be created to allow individual document write events to track new
assets. Once all documents are processed, the list will be converted
back to a simple list for a later stage. When a builder needs to
finalize assets to prepare for publishing, any assets tracked in the
delayed list that are not in the main list will be appended to the
finalized asset list for publishing.

Signed-off-by: James Knight <git@jdknight.me>