flag pre-sphinx-8.1 as non-parallel safe when using plantuml
To support parallel writes for late-produced assets, a multiprocessing
list was introduced [1] for parallel runs. However, this hook only is
available in Sphinx v8.1 or newer. There does not seem to be a trivial
way to extend the capability to older Sphinx versions, so if an older
version of Sphinx is used and plantuml extension is detected, re-flag
the Sphinx extension as not parallel safe for writes.

[1]: f9688c7073d289f5b04bf66574306c7608187699

Signed-off-by: James Knight <git@jdknight.me>