introduce documentation trample detection
This extension has long left responsibility of ensuring unique pages for
Confluence spaces to the user (documenting page updates may pull files
originally on a space into its hierarchy). While the "only new" option
(`confluence_publish_onlynew`) was added to help add some protection,
this only performed a simple ignoring of pages. Since this extension has
evolved over the years, it has more utility methods and cached content
to improve this experience.

This commit introduces a detection of trampling existing documentation
on a space that is not part of the target hierarchy. If a page being
updated is detected to be moved into the published documentation set,
the publish attempt will now fail with a detailed message. Users will
be informed of the failed page to be handled (by either moving the page
or renaming documents).

Users will be able to force the original style of overwriting by
configuring `confluence_publish_trample = True`.

Signed-off-by: James Knight <git@jdknight.me>