correct backref (footnotes/citations) updates in singleconfluence
When the merging of documentation occurs for a `singleconfluence` run,
identifiers are updated to ensure uniqueness. However, for citations and
footnotes, the back references were not updated causing unexpected links
on updated identifiers. To address this, when updating elements with new
identifiers, check for back-references to be updated as well.

Signed-off-by: James Knight <git@jdknight.me>