support epigraphs/pull-quotes with optional attribution
Attribution is not necessarily required for quotes. We used the
attribution tag hint to help distinguish between content that is
typically just indented to quotes that may want a styled left-bar.
While such a check is good for processing block-quote types, we do have
also extra hints for epigraphs and pull-quotes through class markings.
We will take advantage of these markings to force a quote styling if a
epigraph or pull-quote is detected.

Signed-off-by: James Knight <git@jdknight.me>