drop confluence_adv_writer_no_section_cap
This is an advanced/undocumented configuration which was added in case
there was a need to support breaking the sixth header level. Over the
years, it appears there was never a need for such an operating mode.
While cleaning up for a future v3 release, now is a good a time as any
to remove this option.

Signed-off-by: James Knight <git@jdknight.me>