introduce the confluence_cloud option
Provide an option for users to explicitly hint if the configured
Confluence site is a Confluence Cloud instance. This helps force using
certain API v2 calls and use a more appropriate default editor type when
this extension cannot properly determine a Confluence Cloud instance
based on a custom domain site.

Signed-off-by: James Knight <git@jdknight.me>